Update metric's items and properties automatically
The metrics processor used an hard-coded list of items that hasn't been
updated for three years. This now implements fetchers for these items to
update the regularly from WikiData.org

The query to populate the mediawikis has been extended to cover all
wikimedia wikis and not only wikipedias. This should better reflect the
'bad' references.

The list of reference properties is now fetched to include hopefully
the most used reference properties that are not external ids and that
are supposed to be used in references.
There are properties (e.g. P1640, P459) that are not supposed to be used
in references, but are. They are not included and should be fixed on
wikidata.

There will another ticket to create a dashboard to show all the external
ids.

Bug: T209399
Change-Id: I0fd74823a1e6de6bd734b76903e836db4a7eff23